refactor: serde tests
