+++ extensions/common/api/management-api-configuration/src/main/resources/management-api-version.json
@@ -8,7 +8,7 @@
   {
     "version": "5.0.0-beta",
     "urlPath": "/v5beta",
-    "lastUpdated": "2026-07-14T16:00:00Z",
+    "lastUpdated": "2026-07-17T16:00:00Z",
     "maturity": "beta"
   }
 ]
